requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-environ==0.12.0
dotenv==0.9.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ldap3==2.9.1
mysqlclient==2.2.7
phonenumbers==8.12.25
pillow==11.1.0
pyasn1==0.6.1
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
smtpapi==0.4.12
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
